remove code duplication betweenr repeat and while
