+++ openclsim/model/__init__.py
@@ -3,11 +3,10 @@
 from .base_activities import AbstractPluginClass, GenericActivity, PluginActivity
 from .basic_activity import BasicActivity
 from .move_activity import MoveActivity
-from .repeat_activity import RepeatActivity
 from .sequential_activity import SequentialActivity
 from .shift_amount_activity import ShiftAmountActivity
 from .single_run_process import single_run_process
-from .while_activity import WhileActivity
+from .while_activity import WhileActivity, RepeatActivity
 
 __all__ = [
     "AbstractPluginClass",
